src/main/java/com/alexsitiy/ideas/project/validation/EmailCheckValidator.java
package com.alexsitiy.ideas.project.validation;

import com.alexsitiy.ideas.project.repository.UserRepository;
import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor
public class EmailCheckValidator implements ConstraintValidator<EmailCheck, String> {

    private final UserRepository userRepository;

    @Override
    public boolean isValid(String value, ConstraintValidatorContext context) {
        return userRepository.findByUsername(value)
                .isEmpty();
    }
}
